clean root, restructure, rename, decruft
-Rename posts.js to resources.js and move resource-fetching logic
-Clean site metadata html <head> to exclude image, delete large image & old BMX large GIF
-Clear unused imports and images across the project
-Revise Resources
-Fix links in footer
-move most folders from project root into src/
