--- components/layout.tsx
@@ -1,49 +1,0 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/utils.module.css'
-import Link from 'next/link'
-import bg from '../public/images/parkslope.jpg'
-import Providers from './providers'
-import Header from '../sections/Header'
-import Footer from '../sections/Footer'
-
-const name = 'Cooper Reed'
-export const siteTitle = 'Co-Operability'
-
-export default function Layout({
-  children,
-  home,
-}: {
-  children: React.ReactNode
-  home?: boolean // optional prop marked with ?
-}) {
-  return (
-    <div className={styles.container}>
-      <Header />
-      <Providers>
-        <Head>
-          <link rel="icon" type="image/ico" href="/icon.ico" />
-          <meta
-            name="description"
-            content="Cooper's personal website on next.js"
-          />
-          <meta
-            property="og:image"
-            content={`https://og-image.vercel.app/${encodeURI(
-              siteTitle
-            )}.png?theme=light&md=0&fontSize=75px&images=https%3A%2F%2Fassets.zeit.co%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg`}
-          />
-          <meta name="og:title" content={siteTitle} />
-          <meta name="twitter:card" content="summary_large_image" />
-        </Head>
-        <main>{children}</main>
-        {!home && (
-          <div className={styles.backToHome}>
-            <Link href="/">← Back to home</Link>
-          </div>
-        )}
-      </Providers>
-      <Footer />
-    </div>
-  )
-}
