--- resources/linktree.mdx
@@ -1,30 +1,0 @@
----
-title: "🌴Cooper's Linktree"
-date: "2025-04-22"
----
-(all hyperlinked)
-<br />
-**Code:**
-[Resume](https://docs.google.com/document/d/15wtKG9juJMYQOI793LvNYMfmeu7hyErO4xWixkTrSHI/edit?usp=sharing)
- //
-[LinkedIn](https://www.linkedin.com/in/cooperability/)
- //
-[GitHub](https://github.com/cooperability)
- //
-[Google Developer](https://g.dev/cooperability)
-
-<br />
-**Writing:**
-[BlueSky](https://bsky.app/profile/cooperability.com)
- //
-[X](https://x.com/cooperability)
- // 
-[Substack](https://cooperability.substack.com/)
- //
-[Mirror.xyz](https://mirror.xyz/0xAd7f62080c882D575DCd6F5eb29cB9C09B0d4B0D)
-<br />
-
-**Video:**
-[YouTube](https://www.youtube.com/@cooperability)
- //
-[IMDB](https://www.imdb.com/name/nm5343773/)
