+++ src/components/providers.js
@@ -1,0 +1,17 @@
+'use client'
+import { useState, useEffect } from 'react'
+import { ThemeProvider } from 'next-themes'
+
+export default function Providers({ children }) {
+  const [mounted, setMounted] = useState(false)
+
+  useEffect(() => {
+    setMounted(true)
+  }, [])
+
+  if (!mounted) {
+    return <>{children}</>
+  }
+
+  return <ThemeProvider>{children}</ThemeProvider>
+}
