+++ src/pages/api/hello.ts
@@ -1,0 +1,6 @@
+import type { NextApiRequest, NextApiResponse } from 'next'
+
+export default function handler(req: NextApiRequest, res: NextApiResponse) {
+  res.status(200).json({ text: 'Hello' })
+}
+//http://localhost:3000/api/hello
