+++ src/resources/LLMPrompts.mdx
