+++ src/resources/OnlineTools.mdx
@@ -1,0 +1,166 @@
+---
+title: "🛠️Huge List of Online Tools"
+date: "2025-04-23"
+---
+import ToggleDropdown from '../../components/ToggleDropdown';
+
+<ToggleDropdown title="AI">
+[**Google Talk to Books** - limited but _fun_ and _useful_](https://books.google.com/talktobooks/)
+
+[**AI Attack Surface Map** - elegant breakdown of AI Attack Surface](https://danielmiessler.com/blog/the-ai-attack-surface-map-v1-0)
+
+[**ElevenLabs** - free text-to-speech & AI voice generator](https://elevenlabs.io/)
+
+[**Gamma** - AI-powered creation of presentations, documents, and websites](https://gamma.app/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Archives, Repositories, and OSInt">
+[**The Black Vault** - largest declassified government documents archive](https://www.theblackvault.com/documentarchive/)
+
+[**Free Learning List** - free learning resource archive](https://freelearninglist.org/)
+
+[**Shadow Library Uptime Monitor** - for LibGen etc](https://open-slum.org/)
+
+[**PDF Drive** - free PDF archive](https://www.pdfdrive.com/)
+
+[**Transparent California** - California's public records archive](https://transparentcalifornia.com/)
+
+[**Have I Been Pwned** - check if your email has been compromised](https://haveibeenpwned.com/)
+
+[**ICANN Lookup** - check a domain's registry details](https://lookup.icann.org/en)
+
+[**NumLookup** - check a phone number's details for free](https://www.numlookup.com/)
+
+[**Aleph** - OCCRP's OSINT tool for journalists](https://aleph.occrp.org/)
+
+[**Internet Archive Scholar** - search academic papers](https://scholar.archive.org/)
+
+[**100+ Free Data Science Books**](https://www.learndatasci.com/free-data-science-books/)
+
+[**Documenters** - civic participatory media coverage](https://documenters.org/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Blockchain">
+[**L2 Beat** - the state of the layer two ecosystem dashboard](https://l2beat.com/)
+
+[**Stealth Drop** - a stealth address generator](https://stealthdrop.xyz/)
+
+[**Inbox** - E2E encrypted wallet-to-wallet communication](https://about.inb0x.life/)
+
+[**Dustbin** - transform wallet dust into ETH](https://dapp.dustbin.pro/)
+
+[**Sol Incinerator** - transform wallet dust into SOL](https://sol-incinerator.com/)
+
+[**Gashawk** - onchain security, safety, and efficiency](https://www.gashawk.io/)
+
+[**Inclusion Watch** - OFAC Compliance Monitoring](https://www.inclusion.watch/)
+
+[**OpenGSN** - open source gasless transactions; pay gas in any token/fiat](https://opengsn.org/)
+</ToggleDropdown>
+
+
+<ToggleDropdown title="Browser Extensions">
+[** EFF's Privacy Badger** - block ads and trackers](https://www.eff.org/privacybadger)
+
+[**uBlock Origin** - block ads and trackers](https://ublockorigin.com/)
+
+[**Decentraleyes** - block tracking scripts](https://decentraleyes.org/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Content Creation">
+[**ContentIdeas** - curate and analyze trending content](https://www.contentideas.io/)
+
+[**Buffer** - social media scheduling](https://buffer.com/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Cooking">
+[**Just The Recipe** - provide a recipe URL and skip the life stories](https://www.justtherecipe.com/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Data Manipulation">
+[**LibreOffice Online** - free online office suite](https://www.libreoffice.org/download/download-libreoffice/)
+
+[**YouTube Transcript** - get video transcript from a YouTube URL](https://youtubetranscript.com/)
+
+[**PDF OCR** - extract text from a PDF](https://tools.pdf24.org/en/ocr-pdf)
+
+[**iLovePDF** - free PDF tools](https://www.ilovepdf.com/)
+
+[**Unscreen** - remove background from a video](https://www.unscreen.com/)
+
+[**Vocal Remover** - extract vocal stems from a song](https://vocalremover.org/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Digital Rights / Privacy">
+[**Stop junk mail credit card/insurance offers**](https://consumer.ftc.gov/articles/how-stop-junk-mail#:~:text=To%20opt%20out%20permanently%3A%20Go,you've%20started%20the%20process)
+
+[**Submit CCPA complaint**](https://privacy.ca.gov/submit-a-complaint/ccpa-complaints/#:~:text=If%20you%20believe%20a%20business,address%20listed%20on%20the%20form.)
+
+[**Fake Telephone Numbers by Country**](https://fakenumber.org/)
+
+[**Security Education Companion**](https://securityeducationcompanion.org/)
+
+[**EFF's Privacy Badger**](https://www.eff.org/privacybadger)
+
+[**EFF's Street Level Surveillance Catalogue**](https://sls.eff.org/)
+
+[**EFF's Surveillance Self-Defense**](https://ssd.eff.org/)
+
+[**EFF's Why & How to remove Ad ID tracking on iOS and Android**](https://www.eff.org/deeplinks/2022/05/how-disable-ad-id-tracking-ios-and-android-and-why-you-should-do-it-now)
+
+[**EFF's Atlas of Surveillance**](https://atlasofsurveillance.org/)
+</ToggleDropdown>
+
+<ToggleDropdown title="eCommerce">
+[**Import Yeti** - Find any company's suppliers](https://www.importyeti.com/)
+
+[**Keepa** - Amazon price tracker with alerts & historical data](https://keepa.com/#!)
+
+[**TicketsForLess** - ticket reseller](https://www.ticketsforless.com/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Finance">
+[**FreeTaxUSA** - free tax filing, online return prep, E-file Income Tax service](https://www.freetaxusa.com/)
+
+[**IRA Rollover Calculator** - calculate your IRA rollover](https://www.ameriprise.com/financial-news-research/financial-calculators/ira-rollover-calculator)
+
+[**QuiverQuant Corporate Twitter Dashboard** - Twitter engagement metrics](https://www.quiverquant.com/sources/twitter)
+
+[**Layoffs.fyi** - layoff tracker](https://layoffs.fyi/)
+
+[**ADP Calculator** - Payroll, Retirement, Benefits Calculators & tools](https://www.adp.com/resources/tools/calculators.aspx)
+
+</ToggleDropdown>
+
+<ToggleDropdown title="Marketing">
+[**Loading** - graphic asset, icon & animation generator](https://loading.io/)
+</ToggleDropdown>
+
+<ToggleDropdown title="Music">
+[**Synthesia** - AI video generator](https://www.synthesia.io/)
+
+[**Every Noise At Once** - music genre map](http://everynoise.com/engenremap.html)
+
+[**Map of Metal** - metal music genre map](https://mapofmetal.com/)
+
+[**Ishkur's Guide to Electronic Music** - self explanatory](https://music.ishkur.com/)
+</ToggleDropdown>
+
+
+<ToggleDropdown title="Productivity">
+[**How Long To Read** - check how long it takes to read a book](https://howlongtoread.com/)
+</ToggleDropdown>
+
+
+<ToggleDropdown title="Web Development">
+
+[**Web Developer Checklist** - check if your website is compliant](https://webdeveloperchecklist.com/)
+
+[**Screely - Generate/edit Website Mockups**](https://screely.com/)
+
+[**Prestige - in-browser HTTP Client**](https://prestige.dev/#!/doc/browser/master)
+
+[**WebAIM Contrast Checker** - check your website's contrast](https://webaim.org/resources/contrastchecker/)
+
+[**Streamlit** - turn data science scripts into shareable web apps](https://streamlit.io/)
+</ToggleDropdown>
