+++ src/resources/linktree.mdx
@@ -1,0 +1,28 @@
+---
+title: "🌴Cooper's Linktree"
+date: "2025-04-22"
+---
+(all hyperlinked)
+
+**Code:**
+[Resume](https://docs.google.com/document/d/15wtKG9juJMYQOI793LvNYMfmeu7hyErO4xWixkTrSHI/edit?usp=sharing)
+ //
+[LinkedIn](https://www.linkedin.com/in/cooperability/)
+ //
+[GitHub](https://github.com/cooperability)
+ //
+[Google Developer](https://g.dev/cooperability)
+
+**Writing:**
+[BlueSky](https://bsky.app/profile/cooperability.com)
+ //
+[X](https://x.com/cooperability)
+ // 
+[Substack](https://cooperability.substack.com/)
+ //
+[Mirror.xyz](https://mirror.xyz/0xAd7f62080c882D575DCd6F5eb29cB9C09B0d4B0D)
+
+**Video:**
+[YouTube](https://www.youtube.com/@cooperability)
+ //
+[IMDB](https://www.imdb.com/name/nm5343773/)
