+++ tsconfig.json
@@ -25,19 +25,11 @@
     }
   },
   "include": [
-    "src/**/*",
     "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx", 
-    "components/layout.tsx", 
-    "components/providers.js", 
-    "pages/quote.tsx", 
-    "postcss.config.js", 
-    "tailwind.config.js", 
-    "components/date.js", 
-    "components/activeLink.js",
-    "components/demos/**/*"
-, "components/opioid-converter"  ],
+    "src/**/*.ts",
+    "src/**/*.tsx",
+    "next.config.js"
+  ],
   "exclude": [
     "node_modules"
   ]
